+++ dev/com.ibm.ws.wsat.recovery_fat.multi.1/.settings/bndtools.core.prefs
