+++ dev/com.ibm.ws.wsat.recovery_fat.multi.1/fat/src/com/ibm/ws/wsat/fat/FATSuite.java
@@ -14,13 +14,15 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
+import com.ibm.ws.wsat.fat.tests.MultiRecoveryTest1;
+
 import componenttest.rules.repeater.EmptyAction;
 import componenttest.rules.repeater.FeatureReplacementAction;
 import componenttest.rules.repeater.RepeatTests;
 
 @RunWith(Suite.class)
 @SuiteClasses({
-	MultiRecoveryTest.class,
+	MultiRecoveryTest1.class,
 })
 public class FATSuite {
