+++ dev/com.ibm.ws.wsat.recovery_fat.multi.2/.gitignore
@@ -1,0 +1,3 @@
+/fat/src/com/ibm/ws/wsat/fat/tests
+/publish
+/test-applications
