+++ dev/com.ibm.ws.wsat.recovery_fat.multi.2/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>com.ibm.ws.wsat.recovery_fat.multi</name>
+	<name>com.ibm.ws.wsat.recovery_fat.multi.2</name>
 	<comment></comment>
 	<projects>
 	</projects>
